Set version.sbt to 0.8.1-SNAPSHOT (#126)
The greatest published version is 0.8.0. I guess when that was published there was some misconfiguration that prevented the version getting bumped.

Also, a bit confusing, I published a version 0.7.1 with some older dependencies about a week ago. That seems to show up as the "latest" version on the Download badge. So I added a link to Bintray where people can view all the versions.